feat: implement rich_text and related misc objects
