+++ notion_df/object/user.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+from notion_df.utils.mixin import Dictable
+
+
+@dataclass
+class User(Dictable):
+    ...  # TODO
